sweet alert added inside handleDelete block for alert users to sure decition that whether he want delete or not

+++ src/Components/MyTaskCard.jsx
@@ -2,27 +2,45 @@ import React from "react";
 import { Helmet } from "react-helmet-async";
 import { MdDelete, MdEdit } from "react-icons/md";
 import { Link } from "react-router";
+import Swal from "sweetalert2";
 
 const MyTaskCard = ({ myTask, myTasks, setMyTasks }) => {
   const { _id, taskTitle, category, deadLine, budget, email, name } = myTask;
 
   // Handle Delete
   const handleDelete = (id) => {
-    // alert('') swal alert dekhate hobe
-    fetch(`http://localhost:3000/tasks/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount) {
-          // alert('') module dekhe alert daw
-          const remainingMyTasks = myTasks.filter(
-            (myTask) => myTask._id !== id
-          );
-          setMyTasks(remainingMyTasks);
-          console.log("after delete ", data);
-        }
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:3000/tasks/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount) {
+              // alert('') module dekhe alert daw
+              const remainingMyTasks = myTasks.filter(
+                (myTask) => myTask._id !== id
+              );
+              setMyTasks(remainingMyTasks);
+              console.log("after delete ", data);
+            }
+          });
+
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+        });
+      }
+    });
   };
 
   return (
